backend/requirements.txt
django>=4.2
djangorestframework>=3.14
pymongo>=4.5
python-dotenv>=1.0